同じ処理だったから別のusecaseに切り分けた
